fast_api/requirements_fastApi.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
ecdsa==0.19.0
et_xmlfile==2.0.0
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
idna==3.10
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.17
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
